pdfstuff/requirements.txt
tqdm==4.36.1
fpdf==1.7.2
Pillow==6.2.0
